+++ src/components/film-details.js
@@ -1,0 +1,100 @@
+/**
+ * создаёт и возвращает разметку описания фильма в попапе - расширенном описании
+ * @return {String}
+ */
+const getFilmDetailsDescriptionTemplate = () => (
+  `<tr class="film-details__row">
+    <td class="film-details__term">Director</td>
+    <td class="film-details__cell">Anthony Mann</td>
+  </tr>`
+);
+
+/**
+ * создаёт и возвращает разметку кнопок в попапе - расширенном описании
+ * @return {String}
+ */
+const getFilmDetailsControlsTemplate = () => (
+  `<input type="checkbox" class="film-details__control-input visually-hidden" id="watchlist" name="watchlist">
+  <label for="watchlist" class="film-details__control-label film-details__control-label--watchlist">Add to watchlist</label>`
+);
+
+/**
+ * создаёт и возвращает разметку эмоджи в попапе - расширенном описании
+ * @return {String}
+ */
+const getFilmDetailsEmojiTemplate = () => (
+  `<input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="sleeping" checked>
+  <label class="film-details__emoji-label" for="emoji-smile">
+    <img src="./images/emoji/smile.png" width="30" height="30" alt="emoji">
+  </label>`
+);
+
+/**
+ * создаёт и возвращает разметку попапа - расширенного описания
+ * @return {String}
+ */
+export const getFilmDetailsTemplate = () => (
+  `<section class="film-details">
+    <form class="film-details__inner" action="" method="get">
+      <div class="form-details__top-container">
+        <div class="film-details__close">
+          <button class="film-details__close-btn" type="button">close</button>
+        </div>
+        <div class="film-details__info-wrap">
+          <div class="film-details__poster">
+            <img class="film-details__poster-img" src="./images/posters/the-great-flamarion.jpg" alt="">
+
+            <p class="film-details__age">18+</p>
+          </div>
+
+          <div class="film-details__info">
+            <div class="film-details__info-head">
+              <div class="film-details__title-wrap">
+                <h3 class="film-details__title">The Great Flamarion</h3>
+                <p class="film-details__title-original">Original: The Great Flamarion</p>
+              </div>
+
+              <div class="film-details__rating">
+                <p class="film-details__total-rating">8.9</p>
+              </div>
+            </div>
+
+            <table class="film-details__table">
+
+            </table>
+
+            <p class="film-details__film-description">
+              The film opens following a murder at a cabaret in Mexico City in 1936, and then presents the events leading up to it in flashback. The Great Flamarion (Erich von Stroheim) is an arrogant, friendless, and misogynous marksman who displays his trick gunshot act in the vaudeville circuit. His show features a beautiful assistant, Connie (Mary Beth Hughes) and her drunken husband Al (Dan Duryea), Flamarion's other assistant. Flamarion falls in love with Connie, the movie's femme fatale, and is soon manipulated by her into killing her no good husband during one of their acts.
+            </p>
+          </div>
+        </div>
+
+        <section class="film-details__controls">
+
+        </section>
+      </div>
+
+      <div class="form-details__bottom-container">
+        <section class="film-details__comments-wrap">
+          <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">0</span></h3>
+
+          <ul class="film-details__comments-list"></ul>
+
+          <div class="film-details__new-comment">
+            <div for="add-emoji" class="film-details__add-emoji-label">
+              <img src="images/emoji/smile.png" width="55" height="55" alt="emoji">
+            </div>
+
+            <label class="film-details__comment-label">
+              <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment">Great movie!</textarea>
+            </label>
+
+            <div class="film-details__emoji-list">
+
+            </div>
+          </div>
+        </section>
+      </div>
+    </form>
+  </section>`
+);
